Support remote development by specifying extensionKind
This extension does not specify `extensionKind`. It is treated as a `workspace` extension by default meaning it needs to be reinstalled on remote workspaces. See https://code.visualstudio.com/api/advanced-topics/remote-extensions#incorrect-execution-location for more information. This pull request sets `extensionKind` so that it prefers to run as a `ui` extension.